fix type hint for buffer size in STT

+++ livekit-plugins/livekit-plugins-assemblyai/livekit/plugins/assemblyai/stt.py
@@ -59,7 +59,7 @@ class STT(stt.STT):
         end_utterance_silence_threshold: int = 1000,
         http_session: aiohttp.ClientSession | None = None,
         token_expires_in: int = 3600,
-        buffer_size_seconds: int = 0.2,
+        buffer_size_seconds: float = 0.2,
     ):
         super().__init__(streaming_supported=True)
         api_key = api_key or os.environ.get("ASSEMBLYAI_API_KEY")
